move around the files
